WQP-1592 (Drop "unlogged" keyword from all WQP tables)    Removed unlogged from create table sql scripts.    Updated to version v2 for the create changeLog
The approach taken here is that, since wqx.* are regular tables
with no partitions, the setting of the table to logged can be done
separately using a single sql command on existing databases. Going
forward when the liquidbase is used to create the table, the table
will be in the desired state from the beginning.
No need to maintain a separate liquidbase script.